refactor: move hysteresis parameter sweep and analysis to shared engine
Moves the forward and backward simulation sweep and hysteresis calculation logic from handleCheckHysteresis in packages/mcp-server to a reusable, domain-pure service analyzeHysteresis in packages/engine. This removes duplication and maintains shared-engine parity.
