+++ packages/mcp-server/src/handlers/checkHysteresis.ts
@@ -1,7 +1,7 @@
 import { ToolArgs, ToolResult } from '../types/index.js';
 import { z } from 'zod';
 import { createToolResult, parseArgs, parseModelOrThrow, expandModel, cloneExpandedModel, updateMassActionRates } from '../services/engine.js';
-import { simulate, loadEvaluator } from '@bngplayground/engine';
+import { loadEvaluator, analyzeHysteresis } from '@bngplayground/engine';
 import { structureError } from '../services/errors.js';
 
 const checkHysteresisArgsSchema = z.object({
@@ -50,129 +50,22 @@ export async function handleCheckHysteresis(args: ToolArgs): Promise<ToolResult<
             throw new Error('t_end must be a positive number.');
         }
         
-        const forwardValues: number[] = [];
-        const backwardValues: number[] = [];
-        const paramValues: number[] = [];
-        
-        const stepSize = (maxVal - minVal) / (steps - 1);
-        
         await loadEvaluator();
         
-        const obsName = parsedArgs.observable ?? model.observables[0]?.name ?? '';
-        if (!obsName) {
-            throw new Error('No observables available in the model to analyze.');
-        }
-        
-        // Forward sweep - carry state between parameter changes
-        let currentState: Record<string, number> | null = null;
-        let finalForwardState: Record<string, number> = {};
-        
-        for (let i = 0; i < steps; i++) {
-            const paramValue = minVal + i * stepSize;
-            paramValues.push(paramValue);
-            
-            const runModel = cloneExpandedModel(expandedModel);
-            runModel.parameters[parsedArgs.parameter] = paramValue;
-            updateMassActionRates(runModel);
-            
-            // Carry state from previous step
-            if (currentState) {
-                for (const sp of runModel.species) {
-                    if (currentState[sp.name] !== undefined) {
-                        sp.initialConcentration = currentState[sp.name];
-                    }
-                }
-            }
-            
-            const result = await simulate(0, runModel, {
-                method: parsedArgs.method ?? 'ode',
-                t_end: tEnd,
-                n_steps: 50,
-            }, {
-                checkCancelled: () => {},
-                postMessage: () => {},
-            });
-            
-            const lastPoint = result.data[result.data.length - 1];
-            
-            // Save endpoint state for next step
-            currentState = {};
-            for (const key of Object.keys(lastPoint)) {
-                if (key !== 'time') currentState[key] = Number(lastPoint[key]);
-            }
-            
-            // Save final state for backward sweep initialization
-            if (i === steps - 1) {
-                finalForwardState = { ...currentState };
-            }
-            
-            forwardValues.push(Number(lastPoint[obsName] ?? 0));
-        }
-        
-        // Backward sweep - start from LAST forward state (not seed)
-        currentState = finalForwardState;
-        
-        for (let i = 0; i < steps; i++) {
-            const paramValue = minVal + (steps - 1 - i) * stepSize;
-            
-            const runModel = cloneExpandedModel(expandedModel);
-            runModel.parameters[parsedArgs.parameter] = paramValue;
-            updateMassActionRates(runModel);
-            
-            // Use carried state from previous step (or final forward state for first step)
-            for (const sp of runModel.species) {
-                if (currentState[sp.name] !== undefined) {
-                    sp.initialConcentration = currentState[sp.name];
-                }
-            }
-            
-            const result = await simulate(0, runModel, {
-                method: parsedArgs.method ?? 'ode',
-                t_end: tEnd,
-                n_steps: 50,
-            }, {
-                checkCancelled: () => {},
-                postMessage: () => {},
-            });
-            
-            const lastPoint = result.data[result.data.length - 1];
-            
-            // Save endpoint state for next step
-            currentState = {};
-            for (const key of Object.keys(lastPoint)) {
-                if (key !== 'time') currentState[key] = Number(lastPoint[key]);
-            }
-            
-            backwardValues.push(Number(lastPoint[obsName] ?? 0));
-        }
-        
-        // Calculate hysteresis: max difference between forward and backward
-        let maxDiff = 0;
-        let hysteresisRegion: { param: number; diff: number } | null = null;
-        
-        for (let i = 0; i < forwardValues.length; i++) {
-            const diff = Math.abs(forwardValues[i] - backwardValues[i]);
-            if (diff > maxDiff) {
-                maxDiff = diff;
-                hysteresisRegion = { param: paramValues[i], diff };
-            }
-        }
-        
-        const scale = Math.max(...forwardValues.map(Math.abs), 1e-9);
-        const normalizedDiff = maxDiff / scale;
-        
-        const hasHysteresis = normalizedDiff > 0.05;
-        
-        return createToolResult({
-            has_hysteresis: hasHysteresis,
-            hysteresis_magnitude: normalizedDiff,
-            hysteresis_region: hysteresisRegion,
-            forward_curve: paramValues.map((p, i) => ({ param: p, value: forwardValues[i] })),
-            backward_curve: paramValues.map((p, i) => ({ param: p, value: backwardValues[i] })),
-            interpretation: hasHysteresis 
-                ? `Detected hysteresis (${(normalizedDiff * 100).toFixed(1)}% difference). The system shows history-dependent behavior - parameter changes produce different steady-states depending on sweep direction.`
-                : 'No significant hysteresis detected - system behaves reversibly in parameter range.',
+        const result = await analyzeHysteresis({
+            model,
+            expandedModel,
+            parameter: parsedArgs.parameter,
+            sweepRange: parsedArgs.sweep_range as [number, number],
+            steps,
+            observable: parsedArgs.observable ?? '',
+            method: parsedArgs.method ?? 'ode',
+            tEnd,
+            cloneExpandedModel,
+            updateMassActionRates,
         });
+
+        return createToolResult(result);
     } catch (error) {
         const structured = structureError(error instanceof Error ? error : new Error(String(error), { cause: error }));
         return createToolResult(structured);
